requirements.txt
inquirer==3.1.4
pandas==2.0.2
python-dotenv==1.0.0
pytz==2023.3
Requests==2.31.0
XlsxWriter==3.1.2
